feat():level_2实现会议模块大体框架和主要功能
-会议模块主要功能已完成，继续完成通过规则计算日程的功能
-已拆分成不同的计算方法

+++ level2/src/main/java/com/yealink/level1/bean/Staff.java
@@ -32,7 +32,7 @@ public class Staff {
 
     private String email;
 
-    private byte gender;
+    private int gender;
 
     private long createTime;
